Moves MainNavigation to Layout folder
