+++ frontend/src/pages/Root.js
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 
 import classes from './Root.module.css'
-import MainNavigation from '../components/MainNavigation';
+import MainNavigation from '../components/Layout/MainNavigation';
 
 const RootLayout = () => {
     return (<>
